Release 1.02 (#16)
* Added UPX recommendation to README

* README Update

- Added Structure explanation.
- Simplified explanation of execution
- Updated explanation of why Crypter was created

* Further README changes

Corrected a number of typo's

* Fixed broken README prereqs

Also added hyperlinks to point to the correct required versions

* Added link to video demonstration

Added a link in the README to a YouTube video which briefly demonstrates the Crypter build and execution process

* Updated README

Added additional link to the video demonstration, as well as a reference to the issue page for a more in-depth overview of the issues and features currently being addressed

* Implemented working threading of decryption. This is purely for testing.
Currently no decryption is performed, but the process operates as
expected. It simply needs the decryption to be added back in, and to be
tested thoroughly before the feature can be finished and merged

* Completed working implementation of dialog progress bar threading. The
gauge is now threaded, along with the decryption percentage and the number
of files remaining.
TODO: Replace the placeholder code with the actual decryption
functionality

* Resolved issue with maintenance of encrypted and decrypted files list.
Needs cleaning up and a few other fixes need adding

* Fixed one or two GUI bugs. Issue with maintaining encrypted and decrypted
files lists seems to be gone, but this needs confirming. Also need to
investigate potential issue with running under WIn 7

* Added a final pub sub update for the progress

* Completed working decryption threading. Fixed issues with GUI updating
accurately. This isn't quite stable however and should be tested
thoroughly before a release

* Adjusted decryption gauge to set the range as a percentage value out of
100, rather than using the length of the list of encrypted files